[perf] FSDP: speed up no_sync and test communication volume
